requirements.txt
anyio==4.6.0
asgiref==3.8.1
channels==4.1.0
click==8.1.7
Django==5.1.1
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular-sidecar==2024.7.1
gunicorn==23.0.0
h11==0.14.0
httptools==0.6.1
idna==3.10
packaging==24.1
pillow==10.4.0
psycopg2-binary==2.9.9
PyJWT==2.9.0
python-dotenv==1.0.1
PyYAML==6.0.2
sniffio==1.3.1
sqlparse==0.5.1
uvicorn==0.30.6
uvicorn-worker==0.2.0
uvloop==0.20.0
watchfiles==0.24.0
websockets==13.1
